Add support to kotlin.time APIs in the rate limit interceptor.
